chore: update Cargo.toml lints and refactor code for clarity
- Added a lint configuration in Cargo.toml to allow unexpected_cfgs for test checks.
- Refactored the TowerFileSystem instantiation to use a direct struct initialization.
- Updated health checker logic to simplify threshold comparisons.
- Implemented Default trait for HyperHttpClient to streamline instantiation.
- Adjusted BackendUrl tests for improved readability and clarity in assertions.

+++ src/adapters/health_checker.rs
@@ -76,7 +76,7 @@ impl HealthChecker {
                                          target, successes);
                             
                             // If we've reached the threshold, mark as healthy
-                            if successes >= health_config.healthy_threshold as u32
+                            if successes >= health_config.healthy_threshold
                                 && backend_health.status() == HealthStatus::Unhealthy {
                                     
                                 tracing::info!("Backend {} is now HEALTHY (after {} consecutive successes)", 
@@ -119,7 +119,7 @@ impl HealthChecker {
                       target, reason, failures, health_config.unhealthy_threshold);
         
         // Mark as unhealthy if threshold reached and current status is healthy
-        if failures >= health_config.unhealthy_threshold as u32
+        if failures >= health_config.unhealthy_threshold
             && backend_health.status() == HealthStatus::Healthy {
             
             tracing::warn!("Backend {} is now UNHEALTHY (after {} consecutive failures): {}",
